Allow Request State in Autocomplete Query Config
This change allows Request State parameters to be provided in
query configuration. Autocomplete Result Queries will ignore
the current state of the application, filters, sort, etc.,
so in order to allow control of this, static values can be added
to query configuration. So a sort could be applied to the Autocomplete
query independently of the current sort selected by the user.