Added Geo Facet filter support (#260)
Added Geo Facet filter support

Geo facets were already supported, but actually clicking on filters
would cause an API error in App Search.

This is because we derive the filter values for fields based on the
data in the API response for that facet. The problem with that is,
the API response does not return in the response what the original
`center` and `unit` values were. So in order to build a filter from
a facet, we need to hold on to the requested `center` and `unit` values
from the facet request so we can feed them back to the API in filters.

Ex or the original facet API request, which contains a `center` and
`unit`:

```
facets: {
  location: {
    center: "73.102, -78.120",
    type: "range",
    unit: "u",
    ranges: [
      { from: 0, to: 100, name: "Nearby" },
      { from: 100, to: 500, name: "A longer drive" }
    ]
  }
}
```

Ex response:
```
facets: {
  location: [
    {
      type: "range",
      data: [
        {
          to: 100,
          from: 0,
          name: "Nearby",
          count: 0
        },
        {
          to: 500,
          from: 100,
          name: "A longer drive.",
          count: 4
        }
      ]
    }
  ]
}
```

We now inject the originally requested `unit` and `center` values into
the resulting application state (done in AppSearchAPIConnector /
responseAdapter).
```
{
  type: "range",
  field: "location",
  data: [
    {
      count: 0,
      value: {
        center: "73.102, -78.120",
        unit: "u",
        to: 100,
        from: 0,
        name: "Nearby"
      }
    },
    {
      count: 4,
      value: {
        center: "73.102, -78.120",
        unit: "u",
        to: 500,
        from: 100,
        name: "A longer drive."
      }
    }
  ]
}
```

Which we can the pass on to the API in a subsequent request as a
filter:
```
filters: {
  "all": [
    {
      "any": [
        {
          "location": {
            "to": 500,
            "from": 100,
            "unit": "mi",
            "center": "40.037461, -76.30460889999999"
          }
        }
      ]
    }
  ]
}
```

I've additionally updated the demo to include a filter for "Distance" and "Date Established. Distance to show Geo, and Date Established to cover dynamic date ranges, which relates to the bug linked below.

Added support for Geo facet filtering
Updated demo to include Geo facet filtering
Fixed #33 as part of this, because it suffered from the same issue.