Switched SearchProvider to be a functional component with hooks (#469)
This PR changes SearchProvider to use hooks in advance of fixing #464.

In order for hooks to work within the local sandbox environment during
development, we needed to resolve an "invalid hook call" error.

A number of solutions are proposed here:
https://reactjs.org/warnings/invalid-hook-call-warning.html

In our case, the error stemmed from the fact that we link our
dependencies from sandbox to our packages. To resolve that, we
need to also link the react we use in our packages to the react we use
in the sandbox. 😵

I chose to add that linking automatically so that it happens
when running npm run sandbox, to make it seamless for developers.

Since that unfortunately causes issues running tests, as a result, I also added a new check
to tests to inform developers if they still have this linking in place.

Lastly, in our SearchProvider test we were inspect state directly on our instantiated SearchProvider component. That won't work now that it is a functional component using
useState, so we instead need to mock useState in order to access the underlying state.